+++ app/admin/(product-menu)/menu-builder/components/menu-action-bar/index.tsx
@@ -191,6 +191,7 @@ export function MenuActionBar() {
         updateCategory,
         detachCategory,
         detachProductFromCategory,
+        attachProductToCategory,
         attachCategory,
       },
       labels,